feat: migrate team create page into performance folder
