+++ app/(general)/performances/[id]/teams/[teamId]/_components/MemberSessionCard.tsx
