+++ constants/routes.ts
@@ -16,14 +16,12 @@ const ROUTES = {
     LIST: "/performances",
     DETAIL: (id: number) => `/performances/${id}`,
     TEAM: {
+      CREATE: (performanceId: number) => `/performances/${performanceId}/teams/create`,
       LIST: (id: number) => `/performances/${id}/teams`,
-      DETAIL: (performanceId: number, teamId: number) => `/performances/${performanceId}/teams/${teamId}/detail`,
-      EDIT: (performanceId: number, teamId: number) => `/performances/${performanceId}/teams/${teamId}edit`
+      DETAIL: (performanceId: number, teamId: number) => `/performances/${performanceId}/teams/${teamId}`,
+      EDIT: (performanceId: number, teamId: number) => `/performances/${performanceId}/teams/${teamId}/edit`
     },
     EDIT: (id: number) => `/performances/${id}/edit`
-  },
-  TEAM: {
-    CREATE: "/teams/create",
   },
   MEMBER: {
     LIST: "/members",
